requirements.txt
numpy==1.19.5
pillow==8.3.1
tensorflow==2.4.0
opencv-python==4.5.3.56
